% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{distribution_plot}
\alias{distribution_plot}
\title{Distribution Plot}
\usage{
distribution_plot(base_path, location_list, event_list, variable, species_list)
}
\description{
Generates the distribution plot for data visualization
}
