% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{effort_plot}
\alias{effort_plot}
\title{Effort Plot}
\usage{
effort_plot(location, base_path, see_duty_cycle = FALSE, duty_cycle_min = 60)
}
\description{
This function creates the effort/detection plot for the data visualization tab
}
