% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{effort_table}
\alias{effort_table}
\title{effort table}
\usage{
effort_table(location, base_path)
}
\description{
This function aggregates data on which time periods were analyzed.  Output will be in local time zone.
}
