% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata}
\usage{
get_metadata(location, base_path, variable)
}
\description{
Retrieves specific metadata variables 
(Latitude, Longitude, Depth_m, dc, dc_per_hour, tz) 
from csv files in the PAMportal folder.
}
