% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_rds}
\alias{get_rds}
\title{Get RDS Data}
\usage{
get_rds(name, rds_paths, rds_cache_val, update_cache_fn)
}
\description{
Reads in the selected RDS file from the cache
}
\examples{
acou_data <- get_rds(
}
