% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_timezone}
\alias{get_timezone}
\title{Get time zone}
\usage{
get_timezone(location, base_path)
}
\description{
Retrieves time zone for the location.
}
\examples{
get_timezone(location, base_path)
ex. output: "Pacific/Wake"
}
