% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{occr_events}
\alias{occr_events}
\title{Occurrence plot events}
\usage{
occr_events(location, base_path, species_list = c("All"))
}
\description{
Gathers acoustic event data for the Occurrence plot
}
