% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{palette_main}
\alias{palette_main}
\title{Color palettes}
\usage{
palette_main()
}
\arguments{
\item{.data}{A raster or numeric vector providing the data domain for the color scale.}
}
\value{
For `palette_main()`, a named list of hex colors.
}
\description{
PAMPortal color palette. Primarily used for static UI elements.
}
\examples{
palette_main()$green
}
