% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_deep_acoustics}
\alias{plot_deep_acoustics}
\title{DeepAcoustics Plot}
\usage{
plot_deep_acoustics(
  location,
  base_path,
  months_of_interest = c("All"),
  species_of_interest = c("Fin whale", "Blue whale", "Sei whale"),
  grouping = "day",
  load_rds_fn = NULL
)
}
\description{
view DA output compared to PG detections.
}
