% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_hourly_presence}
\alias{plot_hourly_presence}
\title{Hourly presence}
\usage{
plot_hourly_presence(
  location,
  base_path,
  months_of_interest = c("All"),
  species_of_interest = c("All"),
  metric = "Count",
  log_scale = FALSE,
  load_rds_fn = NULL
)
}
\description{
View hourly presence and acoustic intensity by species.  Specify metric as "Count" or "Duration".
}
\examples{
plot_hourly_presence()

}
