% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_measurements}
\alias{plot_measurements}
\title{Measurements}
\usage{
plot_measurements(
  location_list,
  base_path,
  detector_type,
  variables_of_interest,
  species,
  events_of_interest = c("All"),
  load_rds_fn = NULL
)
}
\description{
Compare whistle/click characteristics between multiple events.  Options for detector type are "Whistle & Moan" or "Click".
User can choose any of the variables available for their chosen
}
