% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_occurrence}
\alias{plot_occurrence}
\title{Occurrence plot}
\usage{
plot_occurrence(
  location,
  base_path,
  months_of_interest = c("All"),
  species_of_interest = c("All"),
  environmental_variable = NA,
  show_effort = FALSE,
  load_rds_fn = NULL
)
}
\description{
Plot of number of minutes an animal was detected with option to show total number of 
minutes monitored (effort) for species of interest (default all) and months of interest (default all), 
with environmental variable of interest (options are in the enviro_data list)
}
\examples{
plot_occurrence()

}
