% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{process_acoustic_data}
\alias{process_acoustic_data}
\title{Process Acoustic Data}
\usage{
process_acoustic_data(acou_data)
}
\description{
Process the acoustic data from a specific PAMPal rds file.
}
\examples{
process_acoustic_data(acou_data)
}
