% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{pull_events}
\alias{pull_events}
\title{Pull Events for Effort/Detection Figure}
\usage{
pull_events(location, base_path)
}
\description{
This code pulls all events for a specific location RDS file
}
\examples{
example <- pull_events(location=input$location)
}
