% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{set_colors}
\alias{set_colors}
\title{Set species colors}
\usage{
set_colors(location, base_path)
}
\description{
Creates a map of colors so that species are consistently represented across figures.
}
