% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{sp_annotations}
\alias{sp_annotations}
\title{sp annotation}
\usage{
sp_annotations(location, base_path, duty_cycle_min = 60)
}
\description{
This function annotates date/time information for the effort/detection figure
}
