% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{spectrogram_plotly}
\alias{spectrogram_plotly}
\title{Spectrogram Plot}
\usage{
spectrogram_plotly(
  wave,
  floor = -50,
  background = "#001f3f",
  foreground = "white",
  hover_bgcolor = "white",
  hover_fontcolor = "black",
  overlap = 50,
  zero_padding = 2,
  wl = NULL
)
}
\description{
Function for generating a spectrogram plot
}
