% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{convert_timezone}
\alias{convert_timezone}
\title{Time zone conversion}
\usage{
convert_timezone(df, data_tz, local_tz)
}
\description{
If metadata reflects current time zone in UTC, 
converts an input dataframe to local time zone with day/hour/minute columns
for future use.  Input dataframe MUST have a UTC col.
}
\examples{
convert_timezone(df, data_tz = "UTC", local_tz = "Pacific/Wake")
Output: modified df with local_time, species, callType, duration, day, hour, minute

}
