% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(
  location,
  base_path,
  months_of_interest = c("All"),
  species_of_interest = c("All"),
  load_rds_fn = NULL
)
}
\description{
Retrieves and concatenates all data from the RDS for a 
particular time frame (if specified) into a dataframe.
}
\examples{
get_data(location, base_path, months_of_interest = c(6,7))
output: df with UTC, species, callType, and duration
}
