% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_daylight}
\alias{get_daylight}
\title{Return daylight classification}
\usage{
get_daylight(df, local_tz, location, base_path)
}
\description{
Takes an input dataframe (full grid, converted TZ; must have "day" col) and returns daylight T/F for each row.
}
\examples{
get_daylight(df, local_tz = "Pacific/Wake", location, basepath)

}
