% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_environmental}
\alias{get_environmental}
\title{Get environmental data}
\usage{
get_environmental(location, base_path, months_of_interest = c("All"))
}
\description{
Retrieves environmental df from csv and filters to months of interest.
}
\examples{
get_environmental(location, base_path, c(6,7))

}
