% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_grid}
\alias{get_grid}
\title{Make an expanded grid}
\usage{
get_grid(
  df,
  location,
  base_path,
  months_of_interest = c("All"),
  species_of_interest = c("All"),
  minutes = FALSE
)
}
\description{
Takes an input dataframe (must have columns "day" and "species"), a list of species of interest, 
and specification of minute resolution (T/F) and returns a dataframe with a row for each
day/hour/species/(minute) combination in the given time frame
}
\examples{
get_grid(df, species_list = c("Fin whale", "Blue whale"), minutes = FALSE))

}
