% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_soundmap}
\alias{get_soundmap}
\title{Get sound map}
\usage{
get_soundmap(location, base_path, months_of_interest = c("All"))
}
\description{
Retrieves sound map df from csv and filters to months of interest.
}
\examples{
get_soundmap(location, base_path, c(6,7))

}
