% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\docType{data}
\name{palette_secondary}
\alias{palette_secondary}
\title{secondary color palette used for data viz elements.}
\format{
An object of class \code{character} of length 11.
}
\usage{
palette_secondary
}
\description{
secondary color palette used for data viz elements.
}
\keyword{datasets}
