% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_call_count}
\alias{plot_call_count}
\title{Call count plot}
\usage{
plot_call_count(
  location,
  base_path,
  months_of_interest = c("All"),
  species_of_interest = c("All"),
  environmental_variable = NA,
  log_scale = FALSE,
  load_rds_fn = NULL
)
}
\description{
Plot the number of calls recorded per day (on a regular or logarithmic scale) along with call type.
Option to add environmental data.
}
