% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_call_density}
\alias{plot_call_density}
\title{Call density plot}
\usage{
plot_call_density(
  location,
  base_path,
  months_of_interest = c("All"),
  species_of_interest = c("All"),
  environmental_variable = NA
)
}
\description{
View the relative call density of each species across time.
}
\examples{
plot_call_density()

}
