% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{sp_annotations}
\alias{sp_annotations}
\title{species annotations}
\usage{
sp_annotations(location, base_path)
}
\description{
This function merges event records with effort, changes time zone to local, and pads
}
